# documentation/api-documentation.yml
openapi: "3.0.0"
info:
  version: "3.0.1"
  title: "FemDevs API Documentation"
  description: |-
    The documentation for the FemDevs API.
    Do note that some routes have permission-based requirements that your user must be quilified for in order to access them.
    Check the description for each route in case you are unsure.

    Permissions Overrides:

    The following permissions will bypass all permissions unless its marked otherwise:
    * `Global::Role.Developer`
    * `Global::Role.Administrator`
    * `Global::Role.Owner`
  termsOfService: "https://legal.thefemdevs.com/terms"
  contact:
    name: "FemDevs"
    url: "https://thefemdevs.com"
    email: "support@thefemdevs.com"
  license:
    name: "GNU Affero General Public License v3.0"
    url: "https://www.gnu.org/licenses/agpl-3.0.en.html"
servers:
  - url: "https://api.thefemdevs.com/"
    description: "Production Server"
paths:
...
  /location/coords:
    get:
      operationId: "GetLocationByCoords"
      description: |-
        Gets the location for the given coordinates.
        Returns the location if successful, otherwise returns an error.

        Permissions Required: `Location::Coord`
      summary: "Get Location By Coordinates"
      parameters:
        - in: "header"
          name: "x-coordpair"
          description: "The coordinates to get the location for"
          schema:
            type: string
      responses:
        200:
          description: "The location for the given coordinates"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocationData"
              examples:
                location:
                  value:
                    address:
                      houseNumber: "1600"
                      street: "Pennsylvania Avenue Northwest"
                      city: "Washington"
                      state: "District of Columbia"
                      country: "United States"
                      postalCode: "20500"
                    coords:
                      lat: 38.8976633
                      lng: -77.0365739
                    pluscode: "87G8P9CJ+XQ"
        400:
          description: "Missing coordinates information"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                Missing-Coordinates:
                  value:
                    code: 8
                    message: "Missing Header Parameter(s)"
                Invalid-Coordinates:
                  value:
                    code: 9
                    message: "Invalid Header Parameter(s)"
        401:
          description: "Invalid/Missing API Token"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                Missing-Token:
                  summary: "Missing API Token"
                  value:
                    code: 3
                    message: "No API key provided"
                Invalid-Token:
                  summary: "Invalid API Token"
                  value:
                    code: 4
                    message: "Invalid API key provided"
        403:
          description: "Missing Permissions"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                Missing-Permissions:
                  value:
                    code: 1
                    message: "You are not authorized to access this resource"
                Blacklisted:
                  value:
                    code: 2
                    message: "You are blacklisted from accessing this resource"
        404:
          description: "Error with the given coordinates"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                error:
                  value:
                    code: 22
                    message: "Unable to retreive requested information"
      security:
        - APIToken: []
  /location/pluscode:
    get:
      operationId: "GetLocationByPlusCode"
      description: |-
        Gets the location for the given plus code.
        Returns the location if successful, otherwise returns an error.

        Permissions Required: `Location::Pluscode`
      summary: "Get Location By Plus Code"
      parameters:
        - in: "header"
          name: "x-pluscode"
          description: "The pluscode"
          schema:
            type: string
      responses:
        200:
          description: "The location for the given pluscode"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocationData"
              examples:
                location:
                  value:
                    address:
                      houseNumber: "1600"
                      street: "Pennsylvania Avenue Northwest"
                      city: "Washington"
                      state: "District of Columbia"
                      country: "United States"
                      postalCode: "20500"
                    coords:
                      lat: 38.8976633
                      lng: -77.0365739
                    pluscode: "87G8P9CJ+XQ"
        400:
          description: "Missing pluscode information"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                Missing-Pluscode:
                  value:
                    code: 8
                    message: "Missing Header Parameter(s)"
                Invalid-Pluscode:
                  value:
                    code: 9
                    message: "Invalid Header Parameter(s)"
        401:
          description: "Invalid/Missing API Token"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                Missing-Token:
                  summary: "Missing API Token"
                  value:
                    code: 3
                    message: "No API key provided"
                Invalid-Token:
                  summary: "Invalid API Token"
                  value:
                    code: 4
                    message: "Invalid API key provided"
        403:
          description: "Missing Permissions"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                Missing-Permissions:
                  value:
                    code: 1
                    message: "You are not authorized to access this resource"
                Blacklisted:
                  value:
                    code: 2
                    message: "You are blacklisted from accessing this resource"
        404:
          description: "Error with the given pluscode"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                error:
                  value:
                    code: 22
                    message: "Unable to retreive requested information"
      security:
        - APIToken: []
  /location/address:
    get:
      operationId: "GetLocationByAddress"
      description: |-
        Gets the location for the given address.
        Returns the location if successful, otherwise returns an error.

        Permissions Required: `Location::Address`
      summary: "Get Location By Address"
      parameters:
        - in: "header"
          name: "x-address"
          description: "The address to get the location for"
          schema:
            type: string
      responses:
        200:
          description: "The location for the given address"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocationData"
              examples:
                location:
                  value:
                    address:
                      houseNumber: "1600"
                      street: "Pennsylvania Avenue Northwest"
                      city: "Washington"
                      state: "District of Columbia"
                      country: "United States"
                      postalCode: "20500"
                    coords:
                      lat: 38.8976633
                      lng: -77.0365739
                    pluscode: "87G8P9CJ+XQ"
        400:
          description: "Missing address information"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                Missing-Address:
                  value:
                    code: 8
                    message: "Missing Header Parameter(s)"
                Invalid-Address:
                  value:
                    code: 9
                    message: "Invalid Header Parameter(s)"
        401:
          description: "Invalid/Missing API Token"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                Missing-Token:
                  summary: "Missing API Token"
                  value:
                    code: 3
                    message: "No API key provided"
                Invalid-Token:
                  summary: "Invalid API Token"
                  value:
                    code: 4
                    message: "Invalid API key provided"
        403:
          description: "Missing Permissions"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                Missing-Permissions:
                  value:
                    code: 1
                    message: "You are not authorized to access this resource"
                Blacklisted:
                  value:
                    code: 2
                    message: "You are blacklisted from accessing this resource"
        404:
          description: "Error with the given address"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorData"
              examples:
                error:
                  value:
                    code: 22
                    message: "Unable to retreive requested information"
      security:
        - APIToken: []
...
components:
  securitySchemes:
    APIToken:
      type: "http"
      scheme: "bearer"
      bearerFormat: "JWT"
      description: |-
        JSON Web Token Authentication with information containing your license key

        Permissions:
        - `Global::Data.Read` (0:0:0) - **Unused**
        - `Global::Data.Write` (0:0:1) - **Unused**
        - `Global::Data.Create` (0:0:2) - **Unused**
        - `Global::Data.Delete` (0:0:3) - **Unused**
        - `Global::Token.Read` (0:1:0) - Allows reading of token information
        - `Global::Token.ReadAll` (0:1:1) - Allows reading of all token information
        - `Global::Token.Write` (0:1:2) - **Unused**
        - `Global::Token.Create` (0:1:3) - Allows creating of token information
        - `Global::Token.Delete` (0:1:4) - Allows deleting of token 
        - `Global::User.Read` (0:2:0) - Read user information
        - `Global::User.ReadAll` (0:2:1) - **Unused**
        - `Global::User.Write` (0:2:2) - Write/Update user information
        - `Global::User.Create` (0:2:3) - Create a user profile
        - `Global::User.Delete` (0:2:4) - Delete a user profile
        - `Cryptography::Encrypt` (1:0) - Ability to encrypt data using the `/crypto/enc` endpoint
        - `Cryptography::Decrypt` (1:1) - Ability to decrypt data using the `/crypto/dec` endpoint
        - `Dictionary::Definition` (2:0) - Ability to get the definition of a word using the `/dict/def` endpoint
        - `Dictionary::Synonym` (2:1) - Ability to get the synonyms of a word using the `/dict/syn` endpoint
        - `Dictionary::Antonym` (2:2) - Ability to get the antonyms of a word using the `/dict/ant` endpoint
        - `Facts::Dog` (3:0) - Ability to get a random dog fact using the `/facts/dog` endpoint
        - `Facts::Cat` (3:1) - Ability to get a random cat fact using the `/facts/cat` endpoint
        - `Facts::Random` (3:2) - Ability to get a random fact using the `/facts/random` endpoint
        - `Location::Coord` (4:0) - Ability to get the location of coordinates using the `/location/coords` endpoint
        - `Location::Pluscode` (4:1) - Ability to get the location of a pluscode using the `/location/pluscode` endpoint
        - `Location::Address` (4:2) - Ability to get the location of an address using the `/location/address` endpoint
        - `Phone::Lookup` (5:0) - Ability to get the details of a phone number using the `/phone/details` endpoint
        - `Phone::Validate` (5:1) - **Unused**
        - `Quote::Kanye` (6:0) - Ability to get a random Kanye West quote using the `/quotes/kanye` endpoint
        - `Quote::RonSwanson` (6:1) - Ability to get a random Ron Swanson quote using the `/quotes/ronswanson` endpoint
        - `Quote::Random` (6:2) - Ability to get a random quote using the `/quotes/random` endpoint
        - `Weather::Current` (7:0) - Ability to get the current weather using the `/weather/current` endpoint
        - `Minecraft::Hive.Player` (8:0:0) - Ability to get information about a player on the Hive Minecraft server using the `/minecraft/hive/player` endpoint
...
  schemas:
...
    ErrorData:
      description: |-
        The error data that is returned when an error occurs

        All Error Codes are located in `/functions/errorResponse.js` [GitHub](https://github.com/femdevs/femdev-website/blob/main/functions/errorResponse.js)
      required:
        - code
        - message
      properties:
        code:
          nullable: false
          type: integer
          description: "The error code"
        message:
          nullable: false
          type: string
          description: "The error message"
...
    LocationAddressData:
      description: |-
        The address of the location
      required:
        - full
        - houseNumber
        - street
        - city
        - region
        - country
        - postalCode
      properties:
        full:
          nullable: false
          type: string
          description: "The full address"
        houseNumber:
          nullable: false
          type: string
          description: "The house number"
        street:
          nullable: false
          type: string
          description: "The street"
        city:
          nullable: false
          type: string
          description: "The city"
        region:
          nullable: false
          type: string
          description: "The region"
        country:
          nullable: false
          type: string
          description: "The country"
        postalCode:
          nullable: false
          type: string
          description: "The postal code"
    LocationCoordData:
      description: |-
        The coordinates of the location
      required:
        - lat
        - lng
      properties:
        lat:
          nullable: false
          type: number
          description: "The latitude of the location"
        lng:
          nullable: false
          type: number
          description: "The longitude of the location"
    LocationData:
      description: |-
        The response data for all location endpoints
      required:
        - address
        - coords
        - pluscode
      properties:
        address:
          nullable: false
          type: object
          $ref: "#/components/schemas/LocationAddressData"
          description: "The address of the location"
        coords:
          nullable: false
          type: object
          $ref: "#/components/schemas/LocationCoordData"
          description: "The coordinates of the location"
        pluscode:
          nullable: false
          type: string
          description: "The plus code of the location"
...